test/programmers/feedback/위장/SolutionTest.java
package programmers.feedback.위장;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

class SolutionTest {


    @Test
    void test() {
        Solution solution = new Solution();
        assertEquals(5, solution.solution(new String[][]{
                {"yellow_hat", "headgear"},
                {"blue_sunglasses", "eyewear"},
                {"green_turban", "headgear"}
        }));
        assertEquals(3, solution.solution(new String[][]{
                {"crow_mask", "face"},
                {"blue_sunglasses", "face"},
                {"smoky_makeup", "face"}
        }));

    }
}